Debugging HomaNanoPuArcht

+++ src/internet/model/homa-nanopu-transport.h
@@ -251,7 +251,7 @@ public:
    */
   void DataSendTrace (Ptr<const Packet> p, Ipv4Address srcIp, Ipv4Address dstIp,
                       uint16_t srcPort, uint16_t dstPort, int txMsgId, 
-                      uint16_t pktOffset, uint8_t prio);
+                      uint16_t pktOffset, uint16_t prio);
 
 private:
 
@@ -270,7 +270,7 @@ private:
                  uint16_t, uint8_t> m_dataRecvTrace; //!< Trace of {pkt, srcIp, dstIp, srcPort, dstPort, txMsgId, pktOffset, prio} for arriving DATA packets
   TracedCallback<Ptr<const Packet>, Ipv4Address, Ipv4Address, 
                  uint16_t, uint16_t, int, 
-                 uint16_t, uint8_t> m_dataSendTrace; //!< Trace of {pkt, srcIp, dstIp, srcPort, dstPort, txMsgId, pktOffset, prio} for departing DATA packets
+                 uint16_t, uint16_t> m_dataSendTrace; //!< Trace of {pkt, srcIp, dstIp, srcPort, dstPort, txMsgId, pktOffset, prio} for departing DATA packets
 };   
 
 } // namespace ns3
